pyproject.toml
[tool.poetry]
name = "openstates"
version = "6.18.2"
description = "core infrastructure for the openstates project"
authors = ["James Turk <dev@jamesturk.net>"]
license = "MIT"

[tool.poetry.scripts]
os-update = 'openstates.cli.update:main'
os-initdb = 'openstates.cli.initdb:main'
os-update-computed = 'openstates.cli.update_computed:main'
os-text-extract = 'openstates.cli.text_extract:main'
os-people = 'openstates.cli.people:main'
os-committees = 'openstates.cli.committees:main'
os-us-to-yaml = 'openstates.cli.convert_us:main'
os-scrape = 'openstates.cli.scrape:main'
os-validate = 'openstates.cli.validate:main'

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.8.4"
dj_database_url = "^0.5.0"
Django = "3.2.14"
scrapelib = "^2.0.7"
jsonschema = "^3.2.0"
click = "^8.0"
attrs = "^20.2.0"
PyYAML = "^6.0"
textract = "^1.6"
chardet = "^3.0"
pydantic = "^1.8.2"
spatula = ">=0.8.9,<1.0"
PyJWT = "^2.5.0"
boto3 = "^1.26.61"
us = "^3.1.1"
influxdb-client = "^1.37.0"
google-cloud-logging = "^3.9.0"
google-cloud-secret-manager = "^2.16.4"
google-cloud-storage = "^2.13.0"
google-cloud-pubsub = "^2.19.0"
python-dotenv = "^1.0.0"
django-admin-extra-buttons = "^1.5.7"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
pytest-django = "^3.8.0"
coveralls = "^1.11.1"
flake8 = "^3.7.9"
coverage = {extras = ["toml"], version = "^5.0.4"}
mypy = "^0.900"
types-pytz = "^2021.1.0"
types-click = "^7.1.2"
types-PyYAML = "^5.4.3"
types-requests = "^2.28.11.5"
black = "^22.10.0"

[tool.poetry.group.dev.dependencies]
types-jsonschema = "^4.17.0.3"
types-influxdb-client = "^1.37.0.0"

[tool.coverage.run]
omit = [
  "openstates/data/admin/*",
  "openstates/utils/ansistrm.py",
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
